move entity index communication into ParallelGrid::new
